+++ tests/model_connectors/test_time_series_ets.py
@@ -66,7 +66,7 @@ class TimeSeriesModelConnectorSingleBuildingETSTest(unittest.TestCase):
 
         # now test the spawn connector (independent of the larger geojson translator
         self.time_series = TimeSeriesConnectorETS(sys_params)
-        for b in self.gj.buildings:
+        for b in self.gj.json_loads:
             self.time_series.add_building(b)
 
     def test_timeSeries_init(self):
@@ -91,47 +91,3 @@ class TimeSeriesModelConnectorSingleBuildingETSTest(unittest.TestCase):
 
         results_path = os.path.join(run_path, f"{self.gj.scaffold.project_name}_results")
         self.assertTrue(os.path.join(results_path, 'stdout.log'))
-
-
-# Do not run this case as it takes too long on travis.
-# class SpawnModelConnectorTwoBuildingETSTest(unittest.TestCase):
-#     def setUp(self):
-#         self.data_dir = os.path.join(os.path.dirname(__file__), 'data')
-#         self.output_dir = os.path.join(os.path.dirname(__file__), 'output')
-#
-#         project_name = "spawn_two_building"
-#
-#         if os.path.exists(os.path.join(self.output_dir, project_name)):
-#             shutil.rmtree(os.path.join(self.output_dir, project_name))
-#
-#         # load in the example geojson with a single offie building
-#         filename = os.path.join(self.data_dir, "spawn_geojson_ex2.json")
-#         self.gj = GeoJsonModelicaTranslator.from_geojson(filename)
-#         # use the GeoJson translator to scaffold out the directory
-#         self.gj.scaffold_directory(self.output_dir, project_name)
-#
-#         # load system parameter data
-#         filename = os.path.join(self.data_dir, "spawn_system_params_ex2.json")
-#         sys_params = SystemParameters(filename)
-#
-#         # now test the spawn connector (independent of the larger geojson translator
-#         self.spawn = SpawnConnectorETS(sys_params)
-#         for b in self.gj.buildings:
-#             self.spawn.add_building(b)
-#
-#     def test_spawn_to_modelica_and_run(self):
-#         self.spawn.to_modelica(self.gj.scaffold)
-#
-#         # make sure the model can run using the ModelicaRunner class
-#         mr = ModelicaRunner()
-#         file_to_run = os.path.abspath(
-#             os.path.join(
-#                 self.gj.scaffold.loads_path.files_dir, 'B5a6b99ec37f4de7f94021950', 'SpawnCouplingETS.mo'
-#             )
-#         )
-#         run_path = Path(os.path.abspath(self.gj.scaffold.project_path)).parent
-#         exitcode = mr.run_in_docker(file_to_run, run_path=run_path, project_name=self.gj.scaffold.project_name)
-#         self.assertEqual(0, exitcode)
-#
-#         results_path = os.path.join(run_path, f"{self.gj.scaffold.project_name}_results")
-#         self.assertTrue(os.path.join(results_path, 'stdout.log'))