+++ CHANGELOG.rst
@@ -4,6 +4,8 @@ Change Log
 Version 0.1.1 (Unreleased)
 --------------------------
 * Add ETS data for indirect cooling to system parameters schema
+* Add district system example
+* Update scaffolding to allow for mixed models
 
 Version 0.1.0
 -------------
@@ -11,5 +13,5 @@ Version 0.1.0
 This is the initial release of the package and includes the following functionality:
 
 * Initial implementation of a ModelicaRunner to call a Docker container to run the model.
-* Create a ROM/RC model using Modelica 3.2.x, Modelica Buildings Library 7.0 and TEASER 0.7.2.
+* Create an RC model using Modelica 3.2.x, Modelica Buildings Library 7.0 and TEASER 0.7.2.
 * Create a Spawn-based models which loads an IDF file.